+++ packages/lexical-react/src/shared/AutoFormatterUtils.js
